Resume 7
Challenge

Find a profile image of yourself that is not zoomed into your beautiful face (at least waist up) and make two copies of it.
Crop the copies so that you have one with just shoulders and face and one with only your face.
Use the picture element to show the original image on a laptop-sized screen and bigger, the head and shoulders image on a tablet-sized screen and the face image for mobile.
Include a background image in one of the divs in your resume.
Make use of the CSS background-image properties to ensure it has a backup color property, repeats as many times as you want it to and is in a good position.