Resume 10
Challenge

Add the respective bootstrap classes to your headings, paragraphs and list tags. Refer back to the Typography lesson if needed.

Find all images within your webpage and the respective classes or HTML elements to make it responsive, floating on either your left or right axis, centred in the middle, or with border-radius and border colouring.

Choose one of the following to add your site:
create a form using bootstrap classes and make sure it has a bootstrap button,
implement bootstrap cards and their respective classes,
implement a bootstrap carousel and use its respective classes.

Choose two of the following to add your site:
add bootstrap's jumbotron class to the respective section of your site,
implement bootstrap's modal class and its respective classes,
implement a bootstrap navbar and as a bonus add a breadcrumb to it as a user navigates.

Note: you can choose to do as many of the above options as you wish so that you can explore Bootstrap and its functionality.